@asiehsalehi
-	The math extension of a project scope is restricted. 
In the old version, a local deployed theory was accessible in its project. Now, it needs to be imported in a theorypath within the project.

-	The access of a theorypath is extended. A theorypath can import local/global deployed theories.
In the old version, a theorypath can import only global deployed theories.

-	A theory can import local/global deployed theories.
In the old version, a theory can import only local deployed/non-deployed theories.
o	Forbidding a circular loop in importing the theories is covered. 
o	Checking conflicts between imported theories is covered.

-	Checking conflict of the Axiomatic block of imported theories in a theorypath is done (Other conflict checking (eg. Datatype) was provided before).

svn path=/trunk/TheoryMathsExtensions/; revision=15729