temporary solution to a bug done by Nicolas, it needs to be enhaunced. he bug comes from the fact that org.eventb.core.internal.ast.extensions.maths.OperatorTypingRule.constructPatternType() is using the wrong extension when it calls mediator.makeParametricType()
Debugging calls to
org.eventb.core.ast.extensions.maths.MathExtensionsFactory.getAxiomaticTypeExtension()
which returns different extension instances (a new instance at each call), 
we can see that a first extension is created when importing project MathReal, 
and that other instances are created later on.
These instances get mixed up in formulas, and as the type checker verifies reference equality
of type constructor extensions (not 'equals()', but '==') it results in the syntax error.

svn path=/trunk/TheoryMathsExtensions/; revision=16154