Rewrite WD predicate computation
Use the specialization API from the AST library which fits the need.
However, the implementation of ISpecialization in Rodin 3.1 RC1 contains
two bugs that are worked around by duplicating the code from the AST
library and patching it within the AST extensions plug-in.

When the AST library gets fixed, this kludge should be removed.

svn path=/branches/TheoryMathsExtensions/Rodin3/; revision=16477