Remove the constant BTRUE from AstUtilities
This constant can pose several issues: it is always built with the
default factory and thus shall not be mixed up with formulas built
with other factories.

It is better to publish a method that creates new BTRUE nodes on the
fly and with the appropriate factory.

I have also refactored the conjunctPredicates() methods to eliminate
BTRUE nodes in a much more efficient manner (the previous code was
quadratic).

Finally, add unit tests for the new method.

svn path=/branches/TheoryMathsExtensions/Rodin3/; revision=16541