Fix wrong error about invalid given type in rewrite rule
The method getFreeIdentifiers() has changed since AST 3.0 and now also
returns the given sets that occur indirectly in a formula.  This change
prevented the existing code from returning the proper error message
when a given type is used in a formula.

Also use the appropriate utility method for performing the check, rather
than re-implementing.

This issue was actually detected by a failing test.

svn path=/branches/TheoryMathsExtensions/Rodin3/; revision=16561