Add cache for tactic applications provider
In the proving interface, the list of possible tactic applications is
recalculated every time the goal or an hypothesis is modified. This
caused large slowdowns in some projects (see bug #18).

This adds a generic caching class and uses it to cache the results of
rewrite tactics applications, which was identified by profiling as the
source of the slowdowns (freezes lasting several seconds after each user
action) reported in bug #18. User actions, such as rewriting a proof
term, are now executed in a split second.