Fix failing tests due to change in WD generation
As reported in bug #17, applying a rewrite or inference rule now always
adds a goal to prove the WD condition, even if this condition is true,
which is not expected by some tests that fail due to this additional
subgoal.

As commented by Thai Son Hoang in bug #17, this was is intended to
improve proof reuse: if the WD condition later becomes more complex, the
proof tree will keep the same shape and the user will only have to prove
the new WD condition. Therefore, tests must be adapted to match this new
behavior.