Fix bug #24: ConcurrentModificationException while proving
In a Map, the set returned by keySet() is not a copy of the actual key
set. Therefore, removing a key from the map while iterating on the key
set is not safe and may throw a ConcurrentModificationException in some
circumstances. The removeIf() method of the key set is safe and removes
the elements from the underlying map.