Fix bug #26: replay bug with inference rules
When replaying proofs where an inference or rewrite rule was manually
applied, a proved goal sometimes became uncertain. This was caused by
incompatible formula factories used in the sequent's hypotheses, which
caused the replay to fail and lower the rule's confidence to uncertain.

We now make sure that hypotheses are translated to use the right formula
factory, as suggested in a comment by Thai Son Hoang.