Fix substitution when renaming operators arguments in generated rules
A rewrite rule is generated for each operator of a theory. These rules
need a meta-variable for each of their arguments. Since all rules are
generated in the same block, meta-variables are shared and there may be
conflicts if some operators have arguments that have the same name but
different types. These conflicts are solved by appending _ to the
argument's name, but the substitutions applied to the operator's
definition did not replace the argument by its new name. This fixes it.