Do not add operator arguments' formal name to deployed rule's typing env
When creating deployed rewrite rules, there is no need to add the formal
name of the operator arguments to the typing environment of the rules.
What matters is the name of the meta-variables that are used to rewrite
the operator's definition, and they are added to the typing environment
at the beginning of the method. Moreover, adding both can cause bugs if
there are conflicts (i.e., if a meta-variable and a formal argument have
the same name but a different type).