Add regression test for bug with rewrite rules of operators
There was a bug triggered by having two operators with arguments that
have the same name but different types. It should have been fixed by the
two previous commits (504654e and cd5c00a). Adding these definitions to
the test theory triggers the bug and can be used to check that it is not
reintroduced.