Fix bug #32: Rewrite failure of operator definition
When rewrite rules for operators are generated, a metavariable is added
for each operator argument, which may be renamed if several operators
have arguments with the same name but different types.

Then, when the rule is loaded, we only need to add the metavariables to
the type environment. Adding the original arguments is not only useless,
but will cause bugs when they have been renamed.