Fix bug #39: Type of axiomatic operators not checked
Using parseType() returns a type that is assumed to be well-typed. There
is no check that it is valid in a given type environment. As a
consequence, any syntactically valid type was accepted as return type of
axiomatic operators.

Now, the type is first parsed with parseExpression() and type-checked in
the current type environment. Then, if the expression is valid and is
actually a type, it is accepted.