Add clean method to get a type-checked case expression
Obtaining the type-checked case expression is not trivial, since it
requires building a larger predicate with the inductive argument to be
able to infer the type of the identifiers introduced in the case
expression. This code was duplicated in several places. This puts it in
a method of the corresponding database object.