Only bind unknown identifiers of case expression of recursive operators
When generating the PO for the WD of recursive operators, we have to
bind in an universal quantifier the fresh identifiers of each case.
However, the implementation was binding all identifiers of each case
expression, which could include their type name. Identifiers are now
filtered to only bind those that are not already declared in the type
environment.